+++ tests/conftest.py
@@ -157,3 +157,40 @@ def edit_message_text_response(keyboard: tg_types.InlineKeyboardMarkup) -> dict[
 def edit_message_reply_markup_response(edit_message_text_response: dict[str, typing.Any]) -> dict[str, typing.Any]:
     # unique fixture for test, but really similar to :edit_message_text_response:
     return edit_message_text_response
+
+
+@pytest.fixture
+def edit_message_caption_response() -> dict[str, typing.Any]:
+    return tg_methods.EditMessageCaptionResponse.parse_obj(
+        {
+            'ok': True,
+            'result': tg_types.Message.parse_obj({
+                'chat': tg_types.Chat.parse_obj({
+                    'first_name': 'TestFirstName',
+                    'id': 1234567890,
+                    'last_name': 'TestLastName',
+                    'type': 'private',
+                    'username': 'TestUserName',
+                }),
+                'date': 1686840262,
+                'edit_date': 1686841234,
+                'from_': tg_types.User.parse_obj({
+                    'first_name': 'TestFirstName',
+                    'id': 1234567890,
+                    'is_bot': False,
+                    'language_code': 'ru',
+                    'last_name': 'TestLastName',
+                    'username': 'TestUserName',
+                }),
+                'message_id': 12345,
+                'photo': [{
+                    "file_id": "BQACAgIAAxkBAAMQZJRQU7dV3gHKrckVBQk4NAoy5TsAAvw2AAJq0KlIIuX8ICpuOOwvBA",
+                    "file_unique_id": "AgAD_DYAAmrQqUg",
+                    "file_size": 48,
+                    "width": 10,
+                    "height": 10,
+                }],
+                "caption": "edited caption",
+            }),
+        },
+    ).dict()